Move to using citar for the bibliography
This moves to citar from helm-bibtex for tracking references. In
addition it cleans up the capture templates that are used, relying on
the templates that come from Doom. this is a temp solution to better
understand the actual templates that are needed.

All the code has been moved to the .bak file so it can be restored
later.

Signed-off-by: Brian McGillion <bmg.avoin@gmail.com>